# packages/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "xiaohei-erp",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/api/tenant/usage": {
      "get": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Get Tenant Usage",
        "operationId": "get_tenant_usage_admin_api_tenant_usage_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          },
          {
            "name": "erp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Erp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantUsageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/subscription/redeem-card": {
      "post": {
        "tags": [
          "subscription"
        ],
        "summary": "Redeem Activation Card",
        "operationId": "redeem_activation_card_api_subscription_redeem_card_post",
        "parameters": [
          {
            "name": "erp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Erp Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemActivationCardRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedeemActivationCardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RedeemActivationCardRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 6,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "RedeemActivationCardRequest"
      },
      "RedeemActivationCardResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "plan": {
            "type": "string",
            "title": "Plan"
          },
          "subscription_status": {
            "type": "string",
            "title": "Subscription Status"
          },
          "current_period_ends_at": {
            "type": "string",
            "format": "date-time",
            "title": "Current Period Ends At"
          },
          "added_days": {
            "type": "integer",
            "title": "Added Days"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "plan",
          "subscription_status",
          "current_period_ends_at",
          "added_days"
        ],
        "title": "RedeemActivationCardResponse"
      },
...
      "TenantPlanLimits": {
        "properties": {
          "max_users": {
            "type": "integer",
            "title": "Max Users"
          },
          "max_stores": {
            "type": "integer",
            "title": "Max Stores"
          },
          "max_active_sync_tasks": {
            "type": "integer",
            "title": "Max Active Sync Tasks"
          },
          "max_listings": {
            "type": "integer",
            "title": "Max Listings"
          },
          "autobid_enabled": {
            "type": "boolean",
            "title": "Autobid Enabled"
          },
          "sync_enabled": {
            "type": "boolean",
            "title": "Sync Enabled"
          },
          "extension_enabled": {
            "type": "boolean",
            "title": "Extension Enabled",
            "default": false
          },
          "listing_enabled": {
            "type": "boolean",
            "title": "Listing Enabled",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "max_users",
          "max_stores",
          "max_active_sync_tasks",
          "max_listings",
          "autobid_enabled",
          "sync_enabled"
        ],
        "title": "TenantPlanLimits"
      },
      "TenantRemaining": {
        "properties": {
          "users": {
            "type": "integer",
            "title": "Users"
          },
          "stores": {
            "type": "integer",
            "title": "Stores"
          },
          "active_sync_tasks": {
            "type": "integer",
            "title": "Active Sync Tasks"
          },
          "listings": {
            "type": "integer",
            "title": "Listings"
          }
        },
        "type": "object",
        "required": [
          "users",
          "stores",
          "active_sync_tasks",
          "listings"
        ],
        "title": "TenantRemaining"
      },
...
      "TenantUsage": {
        "properties": {
          "active_users": {
            "type": "integer",
            "title": "Active Users"
          },
          "active_stores": {
            "type": "integer",
            "title": "Active Stores"
          },
          "listings": {
            "type": "integer",
            "title": "Listings"
          },
          "active_tasks": {
            "type": "integer",
            "title": "Active Tasks"
          },
          "active_sync_tasks": {
            "type": "integer",
            "title": "Active Sync Tasks"
          }
        },
        "type": "object",
        "required": [
          "active_users",
          "active_stores",
          "listings",
          "active_tasks",
          "active_sync_tasks"
        ],
        "title": "TenantUsage"
      },
      "TenantUsageResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "plan": {
            "type": "string",
            "title": "Plan"
          },
          "plan_name": {
            "type": "string",
            "title": "Plan Name"
          },
          "subscription_status": {
            "type": "string",
            "title": "Subscription Status"
          },
          "trial_ends_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial Ends At"
          },
          "current_period_ends_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period Ends At"
          },
          "limits": {
            "$ref": "#/components/schemas/TenantPlanLimits"
          },
          "usage": {
            "$ref": "#/components/schemas/TenantUsage"
          },
          "remaining": {
            "$ref": "#/components/schemas/TenantRemaining"
          },
          "features": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Features"
          },
          "is_writable": {
            "type": "boolean",
            "title": "Is Writable"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "plan",
          "plan_name",
          "subscription_status",
          "limits",
          "usage",
          "remaining",
          "features",
          "is_writable",
          "warnings"
        ],
        "title": "TenantUsageResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}